server/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pydantic-settings
sqlalchemy
alembic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
minio 
pyjwt
# Optional dependencies for testing